chore: remove unused code
