+++ source/services/analytics-backoffice/.pylintrc
@@ -1,0 +1,42 @@
+[MASTER]
+
+[MESSAGES CONTROL]
+
+# Only show warnings with the listed confidence levels. Leave empty to show
+# all. Valid levels: HIGH, CONTROL_FLOW, INFERENCE, INFERENCE_FAILURE,
+# UNDEFINED.
+confidence=HIGH,
+           CONTROL_FLOW,
+           INFERENCE,
+           INFERENCE_FAILURE,
+           UNDEFINED
+
+# Disable the message, report, category or checker with the given id(s). You
+# can either give multiple identifiers separated by comma (,) or put this
+# option multiple times (only on the command line, not in the configuration
+# file where it should appear only once). You can also use "--disable=all" to
+# disable everything first and then re-enable specific checks. For example, if
+# you want to run only the similarities checker, you can use "--disable=all
+# --enable=similarities". If you want to run only the classes checker, but have
+# no Warning level messages displayed, use "--disable=all --enable=classes
+# --disable=W".
+disable=global-statement,
+		invalid-name, # snake_case so if name is 'error', not authorized... '_' should be in variable.
+		logging-fstring-interpolation,
+		no-member, # pylint not found constants...
+		no-name-in-module, # pylint not found functions...
+		redefined-outer-name,
+		unexpected-keyword-arg,
+		ungrouped-imports, # We need to call dotenv() function before some imports
+		wrong-import-position, # We need to call dotenv() function before some imports
+
+# Enable the message, report, category or checker with the given id(s). You can
+# either give multiple identifier separated by comma (,) or put this option
+# multiple time (only on the command line, not in the configuration file where
+# it should appear only once). See also the "--disable" option for examples.
+enable=c-extension-no-member
+
+[FORMAT]
+
+# Skip lines with http/https and lines with only strings
+ignore-long-lines=^(\s*\".*)|(\s*f.*)|(.*\s*(# )?<?https?://\S+>?)$
